Fix: properly expand the stack when loading/dumping JSON
When in a native function the stack needs to be expanded manually, this
is especially important when working in loops over objects and arrays.
This should get rid of most mysterious crashes when parsing or dumping json.
They were especially annoying because they would happen or not depending on
whether you did some stuff which expanded the stack before handling JSON
(for example printing the value before serializing it will usually help :)).

Also added tests for these conditions, which caused crashes.